docs(app): miragrate from content layer to content collections
