+++ apps/extension/.gitignore
@@ -37,4 +37,5 @@ yarn-error.log*
 next-env.d.ts
 
 .contentlayer
+.content-collections
 tsconfig.tsbuildinfo
