+++ apps/extension/content/docs/changelog.mdx
@@ -7,8 +7,15 @@ description: All last changes in the project with announcements
 
 ### Cli support
 
-One of the most requested features is the Cli, cause some component have big configuration process, for that we heard you and we made the cli, and with no more config, we use shadcn's config, a big thanks 
-<Link target="_blank" href="https://x.com/dangling_hanma" className="font-semibold" >@Raphael</Link> for that...
+One of the most requested features is the Cli, cause some component have big configuration process, for that we heard you and we made the cli, and with no more config, we use shadcn's config, a big thanks
+
+<Link
+  target="_blank"
+  href="https://x.com/dangling_hanma"
+  className="font-semibold"
+>
+  @Raphael
+</Link>
 
 #### Example
 
@@ -60,8 +67,7 @@ This example showcases the carousel component
 
 ### New team member
 
-This is project is growing up, and this is cool to see from the community, but a lot of parts are missing, so I want you all to welcome a new team member <Link target="_blank" href="https://twitter.com/Soham_Asmi" className="font-semibold" >@Gaurang
-</Link>, together, we feel confident that we can deliver the tools you need to easily build accessible modern UIs faster than ever before.
+This is project is growing up, and this is cool to see from the community, but a lot of parts are missing, so I want you all to welcome a new team member <Link target="_blank" href="https://twitter.com/Soham_Asmi" className="font-semibold" >@Gaurang</Link>, together, we feel confident that we can deliver the tools you need to easily build accessible modern UIs faster than ever before.
 
 ## March 2024 - New Docs
