+++ apps/extension/package.json
@@ -4,16 +4,17 @@
   "private": true,
   "scripts": {
     "dev": "next dev --turbo",
-    "build": "contentlayer2 build && pnpm build:registry && next build && prettier --write \"**/*.{ts,tsx,mdx}\" --cache",
+    "build": "content-collections build && pnpm build:registry && next build && prettier --write \"**/*.{ts,tsx,mdx}\" --cache",
     "start": "next start",
     "lint": "next lint",
     "build:test": "pnpm typecheck && pnpm build:registry && next build && next start -p 3001",
     "build:registry": "npx tsx --tsconfig ./src/script/tsconfig.scripts.json ./src/script/registry-builder.ts",
-    "typecheck": "contentlayer2 build && tsc --noEmit",
-    "watch": "contentlayer2 dev && tsc --noEmit"
+    "typecheck": "content-collections build && tsc --noEmit",
+    "watch": "content-collections dev && tsc --noEmit"
   },
   "dependencies": {
-    "@contentlayer2/source-files": "^0.5.0",
+    "@content-collections/mdx": "^0.2.0",
+    "@content-collections/next": "^0.2.4",
     "@hookform/resolvers": "^3.3.4",
     "@mdx-js/loader": "^3.0.1",
     "@radix-ui/react-accordion": "^1.1.2",
@@ -38,7 +39,6 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "cmdk": "^0.2.1",
-    "contentlayer2": "^0.5.0",
     "date-fns": "^3.3.1",
     "embla-carousel": "^8.0.1",
     "embla-carousel-auto-scroll": "^8.0.1",
@@ -49,7 +49,6 @@
     "mdast-util-toc": "^7.0.0",
     "mini-svg-data-uri": "^1.4.4",
     "next": "^14.1.2",
-    "next-contentlayer2": "^0.5.0",
     "next-themes": "^0.2.1",
     "posthog-js": "^1.116.5",
     "react": "^18",
@@ -60,7 +59,7 @@
     "react-otp-input": "^3.1.1",
     "react-resizable-panels": "^2.0.12",
     "react-wrap-balancer": "^1.1.0",
-    "rehype-pretty-code": "^0.13.0",
+    "rehype-pretty-code": "^0.13.2",
     "rehype-slug": "^6.0.0",
     "remark": "^15.0.1",
     "remark-code-import": "^1.2.0",
@@ -80,10 +79,13 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@content-collections/cli": "^0.1.6",
+    "@content-collections/core": "^0.8.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
+    "concurrently": "^9.1.2",
     "eslint": "^8",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
